[FIX] privacy_consent: Markup controller (partner) name before rendering it
As we now use t-out in frontend templates we cannot inject directly html
content. This was possible with t-raw that is deprecated.

In this commit we fix that issue by using Markup in order to create valid
html content based on partner html-ified name and use it in template.

Other solution would be to have display_name of partner being automatically
updated depending on context keys, but as it is a stored field it cannot be
done.